+++ Plugin/EnvironmentPlugin/ProjectDescriptionHelpers/ProjectEnvironment.swift
@@ -7,7 +7,6 @@ public struct ProjectEnvironment {
     public let deploymentTarget: DeploymentTarget
     public let platform: Platform
     public let baseSetting: SettingsDictionary
-    public let isCI: Bool
 }
 
 public let env = ProjectEnvironment(
@@ -15,6 +14,5 @@ public let env = ProjectEnvironment(
     organizationName: "",
     deploymentTarget: .iOS(targetVersion: "16.0", devices: [.iphone, .ipad]),
     platform: .iOS,
-    baseSetting: [:],
-    isCI: (ProcessInfo.processInfo.environment["TUIST_CI"] ?? "0") == "1" ? true : false
+    baseSetting: [:]
 )
