+++ Tuist/ProjectDescriptionHelpers/Project/Project+makeModule.swift
@@ -33,7 +33,7 @@ public extension Project {
         additionalFiles: [FileElement] = [],
         configurations: [Configuration] = []
     ) -> Project {
-        let scripts: [TargetScript] = env.isCI ? [] : [.swiftLint]
+        let scripts: [TargetScript] = generateEnvironment.scripts
         let ldFlagsSettings: SettingsDictionary = product == .framework ?
         ["OTHER_LDFLAGS": .string("$(inherited) -all_load")] :
         ["OTHER_LDFLAGS": .string("$(inherited)")]
@@ -192,41 +192,6 @@ public extension Project {
             settings: settings,
             targets: allTargets,
             schemes: schemes
-        )
-    }
-}
-
-extension Scheme {
-    static func makeScheme(target: ConfigurationName, name: String) -> Scheme {
-        return Scheme(
-            name: name,
-            shared: true,
-            buildAction: .buildAction(targets: ["\(name)"]),
-            testAction: .targets(
-                ["\(name)Tests"],
-                configuration: target,
-                options: .options(coverage: true, codeCoverageTargets: ["\(name)"])
-            ),
-            runAction: .runAction(configuration: target),
-            archiveAction: .archiveAction(configuration: target),
-            profileAction: .profileAction(configuration: target),
-            analyzeAction: .analyzeAction(configuration: target)
-        )
-    }
-    static func makeDemoScheme(target: ConfigurationName, name: String) -> Scheme {
-        return Scheme(
-            name: name,
-            shared: true,
-            buildAction: .buildAction(targets: ["\(name)DemoApp"]),
-            testAction: .targets(
-                ["\(name)Tests"],
-                configuration: target,
-                options: .options(coverage: true, codeCoverageTargets: ["\(name)DemoApp"])
-            ),
-            runAction: .runAction(configuration: target),
-            archiveAction: .archiveAction(configuration: target),
-            profileAction: .profileAction(configuration: target),
-            analyzeAction: .analyzeAction(configuration: target)
         )
     }
 }
